refactor: structural cleanups from audit follow-up
- migrate the last 7 files from @/lib/supabase-client to the canonical
  @/integrations/supabase/client import
- delete src/legacy/ (verified unreferenced in two consecutive audits)
- secure-flex-api: FLEX_API_BASE_URL now configurable via env
- import-users: bump deno std 0.182.0 -> 0.224.0 (matches the rest)
- vitest: coverage thresholds for the new shared flexFetch and
  whatsappQuota helpers
