+++ src/components/feedback/AdminPanel.tsx
@@ -39,7 +39,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import { supabase } from "@/lib/supabase-client";
+import { supabase } from "@/integrations/supabase/client";
 import { useToast } from "@/hooks/use-toast";
 import { ExternalLink, Trash2, Edit, CheckCircle, Eye } from "lucide-react";
 import { format } from "date-fns";
