--- src/legacy/components/Layout.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/components/layout/Layout";
