+++ supabase/functions/import-users/index.ts
@@ -1,7 +1,7 @@
 
 import { serve } from "https://deno.land/std@0.224.0/http/server.ts";
 import { createClient } from "npm:@supabase/supabase-js@2";
-import { parse } from "https://deno.land/std@0.182.0/csv/parse.ts";
+import { parse } from "https://deno.land/std@0.224.0/csv/parse.ts";
 
 const corsHeaders = {
   'Access-Control-Allow-Origin': '*',
